pyproject.toml
[tool.poetry]
name = "audio-transformers"
version = "0.1.0"
description = "Audio data transformations library with a command-line interface."
authors = ["Stepan Anokhin <stepan.anokhin@gmail.com>"]
include = ["README.md"]
readme = "README.md"
homepage = "https://github.com/stepan-anokhin/audio-transformers"
repository = "https://github.com/stepan-anokhin/audio-transformers"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.0.1"
scipy = "^1.14.0"
ffmpegio = "^0.10.0.post0"
pytest-cov = "^5.0.0"
humanize = "^4.10.0"
tabulate = "^0.9.0"
tqdm = "^4.66.5"
pyyaml = "^6.0.2"
termcolor = "^2.4.0"
fire = "^0.6.0"
requests = "^2.32.3"
dacite = "^1.8.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"
flake8 = "^7.1.1"

[tool.poetry.scripts]
audio = 'audio_transformers.cli.root_handler:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
